USB keyboard code re-written
The code for dealing with the USB keyboard reading has been re-written to allow roll-over.

The original code just read the first entry in the USB keyboard report but I found that on a keyboard with proper key switches key presses would be missed due to a previous key not being released in time.

Reading multiple keys using the tinyusb library is made more complex by the library not returning how many keys are currently pressed and the report also doesn't hold them in chronological order or even packed into the 6 keycode length table. So, this code has to take care of all of this.

This code still isn't able to report multiple keys pressed at the same time to the QL but it will be able to report multiple sequensial presses.